requirements.txt
fastapi
uvicorn[standard]
aiogram>=3.0.0
python-dotenv
